requirements.txt
Flask==3.0.1
python-docx==1.1.0